test: move to inspect-multiple-clients.test.ts and wire socket close/error to reject
inspect.test.ts has pre-existing websocket describe-block failures in
the gate/local environment that mask this test's fail-before/pass-after
signal. The test/cli/inspect/ directory already splits by feature area
(test-reporter, HTTPServerAgent, BunFrontendDevServer), so the
multi-client test gets its own file.

Also wire close/error on both sockets to reject every in-flight await
so a future regression that crashes the inspectee fails fast with a
diagnostic instead of hanging to the suite timeout.
